requirements.txt
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
lightgbm>=4.0
pulp>=2.7
pyyaml>=6.0
streamlit>=1.30
torch>=2.1
transformers==4.57.3
tqdm>=4.66
